src/main/java/com/acme/auction/Item.java
package com.acme.auction;

import java.util.Comparator;
import java.util.Objects;

public class Item implements Comparable<Item> {

    private final String id;

    private final String description;

    public Item(String id, String description) {
        this.id = id;
        this.description = description;
    }

    public String getId() {
        return id;
    }

    public String getDescription() {
        return description;
    }

    @Override
    public int hashCode() {
        int hash = 7;
        hash = 19 * hash + Objects.hashCode(this.id);
        return hash;
    }

    @Override
    public boolean equals(Object obj) {
        if (this == obj) {
            return true;
        }
        if (obj == null) {
            return false;
        }
        if (getClass() != obj.getClass()) {
            return false;
        }
        final Item other = (Item) obj;
        if (!Objects.equals(this.id, other.id)) {
            return false;
        }
        return true;
    }

    @Override
    public int compareTo(Item o) {
        return Comparator.comparing(Item::getId).compare(this, o);
    }


    public static class Builder {

        private Builder() {
        }

        public static Builder create() {
            return new Builder();
        }

        public Builder withId(String id) {
            this.id = id;
            return this;
        }

        public Builder withDescription(String description) {
            this.description = description;
            return this;
        }

        private String id;

        private String description;

        public Item build() {
            
            
            return new Item(id, description);
        }
    }

}
